changes in req and res pattern

+++ api/models/note.go
@@ -1,7 +1,7 @@
 package models
 
 type Note struct {
-	ID      uint64 `json:"id"`
+	ID      uint32 `json:"id"`
 	Content string `json:"content"`
 	Author  string `json:"author"`
 }
